+++ bubbleSim/particle.cpp
@@ -414,7 +414,7 @@ numType ParticleGenerator::generateNParticlesInSphere(
 ParticleCollection::ParticleCollection(unsigned int t_particleCountTrue,
                                        unsigned int t_particleCountFalse,
                                        bool t_bubbleIsTrueVacuum,
-                                       std::uint32_t& t_buffer_flags,
+                                       std::uint64_t& t_buffer_flags,
                                        cl::Context& cl_context) {
   // Set up random number generator
   int openCLerrNum;
@@ -518,6 +518,7 @@ ParticleCollection::ParticleCollection(unsigned int t_particleCountTrue,
                  m_particleCountTotal * sizeof(cl_uint),
                  m_particle_collision_cell_index.data(), &openCLerrNum);
   t_buffer_flags |= PARTICLE_COLLISION_CELL_IDX_BUFFER;
+  
   // Data reserve
   m_interacted_bubble_false_state =
       std::vector<int8_t>(m_particleCountTotal, 0);
